Move test macros to a new header and simplify macros
