+++ test/common/test_common.cpp
@@ -48,6 +48,7 @@
 #include <pcl/point_cloud.h>
 
 #include <pcl/common/centroid.h>
+#include <pcl/test/pcl_macros.h>
 
 using namespace pcl;
 
@@ -308,7 +309,7 @@ TEST (PCL, PointTypes)
 
 template <typename T> class XYZPointTypesTest : public ::testing::Test { };
 using XYZPointTypes = ::testing::Types<BOOST_PP_SEQ_ENUM(PCL_XYZ_POINT_TYPES)>;
-PCL_TYPED_TEST_SUITE(XYZPointTypesTest, XYZPointTypes);
+TYPED_TEST_SUITE(XYZPointTypesTest, XYZPointTypes);
 TYPED_TEST(XYZPointTypesTest, GetVectorXfMap)
 {
   TypeParam pt;
@@ -329,7 +330,7 @@ TYPED_TEST(XYZPointTypesTest, GetArrayXfMap)
 
 template <typename T> class NormalPointTypesTest : public ::testing::Test { };
 using NormalPointTypes = ::testing::Types<BOOST_PP_SEQ_ENUM(PCL_NORMAL_POINT_TYPES)>;
-PCL_TYPED_TEST_SUITE(NormalPointTypesTest, NormalPointTypes);
+TYPED_TEST_SUITE(NormalPointTypesTest, NormalPointTypes);
 TYPED_TEST(NormalPointTypesTest, GetNormalVectorXfMap)
 {
   TypeParam pt;
@@ -341,7 +342,7 @@ TYPED_TEST(NormalPointTypesTest, GetNormalVectorXfMap)
 
 template <typename T> class RGBPointTypesTest : public ::testing::Test { };
 using RGBPointTypes = ::testing::Types<BOOST_PP_SEQ_ENUM(PCL_RGB_POINT_TYPES)>;
-PCL_TYPED_TEST_SUITE(RGBPointTypesTest, RGBPointTypes);
+TYPED_TEST_SUITE(RGBPointTypesTest, RGBPointTypes);
 TYPED_TEST(RGBPointTypesTest, GetRGBVectorXi)
 {
   TypeParam pt; pt.r = 1; pt.g = 2; pt.b = 3; pt.a = 4;
