+++ test/io/test_io.cpp
@@ -48,6 +48,7 @@
 #include <pcl/io/ply_io.h>
 #include <pcl/io/ascii_io.h>
 #include <pcl/io/obj_io.h>
+#include <pcl/test/pcl_macros.h>
 #include <fstream>
 #include <locale>
 #include <stdexcept>
@@ -1319,7 +1320,7 @@ TEST (PCL, Locale)
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename T> class AutoIOTest : public testing::Test { };
 using PCLXyzNormalPointTypes = ::testing::Types<BOOST_PP_SEQ_ENUM (PCL_XYZ_POINT_TYPES PCL_NORMAL_POINT_TYPES)>;
-PCL_TYPED_TEST_SUITE (AutoIOTest, PCLXyzNormalPointTypes);
+TYPED_TEST_SUITE (AutoIOTest, PCLXyzNormalPointTypes);
 TYPED_TEST (AutoIOTest, AutoLoadCloudFiles)
 {
   PointCloud<TypeParam> cloud;
