+++ test/octree/test_octree_iterator.cpp
@@ -35,13 +35,14 @@
  *
  *
  */
-#include <gtest/gtest.h>
 #include <pcl/pcl_macros.h>
 #include <pcl/octree/octree_pointcloud_adjacency.h>
 #include <pcl/octree/octree_base.h>
 #include <pcl/octree/octree_iterator.h>
 #include <pcl/common/projection_matrix.h>
 #include <pcl/point_types.h>
+#include <gtest/gtest.h>
+#include <pcl/test/pcl_macros.h>
 
 using pcl::octree::OctreeBase;
 using pcl::octree::OctreeIteratorBase;
@@ -155,7 +156,7 @@ using OctreeIteratorTypes = testing::Types
          OctreeLeafNodeDepthFirstIterator<OctreeBase<int> >,
          OctreeFixedDepthIterator<OctreeBase<int> >,
          OctreeLeafNodeBreadthFirstIterator<OctreeBase<int> > >;
-PCL_TYPED_TEST_SUITE (OctreeIteratorTest, OctreeIteratorTypes);
+TYPED_TEST_SUITE (OctreeIteratorTest, OctreeIteratorTypes);
 
 TYPED_TEST (OctreeIteratorTest, CopyConstructor)
 {
